Keep existing id_token if not in token result
Returning an ID token from requests to the token endpoint using the
`refresh_token` grant is optional. Thus also applying the ID token to
the user data structure must be optional if no new `id_token` value is
returned. This change implements just that in a very similiar way as how
the existing `refresh_token` value is kept.

Fixes: #1236